requirements.txt
streamlit
python-chess
stockfish
pillow
cairosvg
